Spatial-awareness/pyproject.toml
[project]
name = "spatial-awareness"
version = "0.1.0"
description = "3D factory reconstruction from smartphone video/images"
requires-python = ">=3.10"

dependencies = [
    "numpy>=1.26",
    "opencv-python>=4.10",
    "opencv-contrib-python>=4.10",
    "Pillow>=10.0",
    "open3d>=0.18",
    "trimesh>=4.0",
    "plyfile>=1.0",
    "typer>=0.12",
    "rich>=13.7",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
    "sqlalchemy>=2.0",
]
gpu = [
    "gsplat>=1.0",
]
semantic = [
    "anthropic>=0.39",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "httpx>=0.27",
    "ruff>=0.3",
]

[project.scripts]
spatial = "spatial.cli:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
